Add GET /user/reviews, POST /user/review/{id} (must be edited)

+++ Back/src/main/java/com/donggukReview/donggukReview/entity/Review.java
@@ -36,7 +36,7 @@ public class Review {
     private String reviewRatingsTotal;
 
     @Column(name = "review_recommended")
-    private Long reviewRecommended;
+    private Integer reviewRecommended;
 
     @Column(name = "user_id", unique = true, nullable = false)
     private Long userId;
